netops: Use the size_t, Luke

+++ src/netops.h
@@ -12,18 +12,18 @@ typedef unsigned int GIT_SOCKET;
 
 typedef struct gitno_buffer {
 	char *data;
-	unsigned int len;
-	unsigned int offset;
+	size_t len;
+	size_t offset;
 	GIT_SOCKET fd;
 } gitno_buffer;
 
 void gitno_buffer_setup(gitno_buffer *buf, char *data, unsigned int len, int fd);
 int gitno_recv(gitno_buffer *buf);
 void gitno_consume(gitno_buffer *buf, const char *ptr);
-void gitno_consume_n(gitno_buffer *buf, unsigned int cons);
+void gitno_consume_n(gitno_buffer *buf, size_t cons);
 
 int gitno_connect(const char *host, const char *port);
-int gitno_send(int s, const char *msg, int len, int flags);
+int gitno_send(int s, const char *msg, size_t len, int flags);
 int gitno_select_in(gitno_buffer *buf, long int sec, long int usec);
 
 #endif
